fix(ci): align compose tests with current semantics
